T-9035: Fixed XSS in spam cleaner and tinymce course files
Reviewed-on: https://review.totaralms.com/798
Tested-by: Totara Hudson <totara-hudson@catalyst.net.nz>

+++ admin/report/spamcleaner/index.php
@@ -35,21 +35,22 @@ require_once($CFG->dirroot . '/lib/pear/HTML/AJAX/JSON.php'); // required for PH
 
 require_js(array('yui_dom-event', 'yui_connection', 'yui_json'));
 
-$keyword = optional_param('keyword', '', PARAM_RAW);
-$autodetect = optional_param('autodetect', '', PARAM_RAW);
-$del = optional_param('del', '', PARAM_RAW);
-$delall = optional_param('delall', '', PARAM_RAW);
-$ignore = optional_param('ignore', '', PARAM_RAW);
-$reset = optional_param('reset', '', PARAM_RAW);
-$id = optional_param('id', '', PARAM_INT);
+$keyword = optional_param('keyword', '', PARAM_ALPHA);
+$autodetect = optional_param('autodetect', '', PARAM_ALPHA);
+$del = optional_param('del', '', PARAM_ALPHA);
+$delall = optional_param('delall', '', PARAM_ALPHA);
+$ignore = optional_param('ignore', '', PARAM_ALPHA);
+$reset = optional_param('reset', '', PARAM_ALPHA);
+$id = optional_param('id', 0, PARAM_INT);
 
 require_login();
+require_capability('moodle/site:config', get_context_instance(CONTEXT_SYSTEM));
 admin_externalpage_setup('reportspamcleaner');
 
-// Implement some AJAX calls 
+// Implement some AJAX calls
 
 // Delete one user
-if (!empty($del) && confirm_sesskey() && ($id != $USER->id)) {
+if (!empty($del) && confirm_sesskey() && !empty($id) && ($id != $USER->id)) {
     if (isset($SESSION->users_result[$id])) {
         $user = $SESSION->users_result[$id];
         if (delete_user($user)) {
@@ -79,7 +80,7 @@ if (!empty($delall) && confirm_sesskey()) {
     exit;
 }
 
-if (!empty($ignore)) {
+if (!empty($ignore) && !empty($id)) {
     unset($SESSION->users_result[$id]);
     echo json_encode(true);
     exit;
@@ -102,7 +103,7 @@ print_box_start();     // The forms section at the top
 
 <form method="post" action="index.php">
   <div>
-    <input type="text" name="keyword" id="keyword_el" value="<?php p($keyword) ?>" /> 
+    <input type="text" name="keyword" id="keyword_el" value="<?php p($keyword) ?>" />
     <input type="hidden" name="sesskey" value="<?php echo sesskey();?>" />
     <input type="submit" value="<?php echo get_string('spamsearch', 'report_spamcleaner')?>" />
   </div>
@@ -121,7 +122,7 @@ print_box_start();     // The forms section at the top
 </div>
 
 <?php
-print_box_end(); 
+print_box_end();
 
 echo '<div id="result" class="mdl-align">';
 
@@ -143,12 +144,12 @@ echo '</div>';
 /////////////////////////////////////////////////////////////////////////////////
 
 
-///  Functions 
+///  Functions
 
 
 function search_spammers($keywords) {
 
-    global $CFG, $USER; 
+    global $CFG, $USER;
 
     if (!is_array($keywords)) {
         $keywords = array($keywords);    // Make it into an array
@@ -307,7 +308,7 @@ function init() {
     YAHOO.util.Event.addListener("removeall_btn", "click", function(){
         var yes = confirm('<?php echo get_string('spamdeleteallconfirm', 'report_spamcleaner');?>');
         if(yes){
-            var cObj = YAHOO.util.Connect.asyncRequest('POST', '<?php echo me();?>?delall=yes&sesskey=<?php echo $sesskey;?>', delall_cb); 
+            var cObj = YAHOO.util.Connect.asyncRequest('POST', '<?php echo me();?>?delall=yes&sesskey=<?php echo $sesskey;?>', delall_cb);
         }
     });
 }
@@ -355,18 +356,18 @@ function del_user(obj, id) {
     var yes = confirm('<?php echo get_string('spamdeleteconfirm', 'report_spamcleaner');?>');
     if(yes){
         row = obj;
-        var cObj = YAHOO.util.Connect.asyncRequest('POST', '<?php echo me();?>?del=yes&sesskey=<?php echo $sesskey;?>&id='+id, del_cb); 
+        var cObj = YAHOO.util.Connect.asyncRequest('POST', '<?php echo me();?>?del=yes&sesskey=<?php echo $sesskey;?>&id='+id, del_cb);
     }
 }
 function ignore_user(obj, id) {
     row = obj;
-    var cObj = YAHOO.util.Connect.asyncRequest('POST', '<?php echo me();?>?ignore=yes&sesskey=<?php echo $sesskey;?>&id='+id, ignore_cb); 
+    var cObj = YAHOO.util.Connect.asyncRequest('POST', '<?php echo me();?>?ignore=yes&sesskey=<?php echo $sesskey;?>&id='+id, ignore_cb);
 }
 YAHOO.util.Event.onDOMReady(init);
 //]]>
 </script>
 
-<?php 
+<?php
 }
 
 admin_externalpage_print_footer();
